+++ apps/testapp/src/main.tsx
@@ -1,21 +1,22 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import { ThemeProvider, createTheme, CssBaseline } from '@mui/material'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { GraviolaProvider } from './provider/GraviolaProvider'
-import './index.css'
-import App from './App.tsx'
-import { schema } from './schema.ts'
-import { allRenderers } from './provider/config.ts'
-
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { ThemeProvider, createTheme, CssBaseline } from "@mui/material";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { GraviolaProvider } from "./provider/GraviolaProvider";
+import { bringDefinitionToTop } from "@graviola/json-schema-utils";
+import "./index.css";
+import App from "./App.tsx";
+import { schema } from "./schema.ts";
+import { allRenderers } from "./provider/config.ts";
+import { generateDefaultUISchema } from "@graviola/edb-ui-utils";
 // Create a theme instance
 const theme = createTheme({
   palette: {
     primary: {
-      main: '#1976d2',
+      main: "#1976d2",
     },
     secondary: {
-      main: '#dc004e',
+      main: "#dc004e",
     },
   },
 });
@@ -30,43 +31,69 @@ const queryClient = new QueryClient({
   },
 });
 
+const uischemata = {
+  Item: generateDefaultUISchema(
+    bringDefinitionToTop(schema as any, "Item") as any,
+    {
+      scopeOverride: {
+        "#/properties/tags": {
+          type: "Control",
+          scope: "#/properties/tags",
+          options: {
+            chips: true,
+            dropdown: true,
+          },
+        },
+      },
+    },
+  ),
+};
 
-createRoot(document.getElementById('root')!).render(
+createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <QueryClientProvider client={queryClient}>
       <ThemeProvider theme={theme}>
         <CssBaseline />
-        <GraviolaProvider 
-          apiBaseUrl="http://localhost:3000/api/datastore" 
-          schema={schema as any} 
+        <GraviolaProvider
+          apiBaseUrl="http://localhost:7898/query"
+          schema={schema as any}
           renderers={allRenderers}
-          baseIRI={"https://ausleihe.freie-theater-sachsen.de/"}
-          entityBaseIRI={"https://entities.freie-theater-sachsen.de/"}
+          baseIRI={"http://www.example.org/"}
+          entityBaseIRI={"http://www.example.org/Item/"}
           primaryFields={{
-            "Category": {
-              "label": "name",
-              "description": "description"
+            Category: {
+              label: "name",
+              description: "description",
             },
-            "Item": {
-              "label": "name",
-              "description": "description",
-              "image": "photos"
-            }
+            Item: {
+              label: "name",
+              description: "description",
+              image: "photos",
+            },
+            Tag: {
+              label: "name",
+              description: "description",
+              image: "image",
+            },
           }}
           typeNameLabelMap={{
             Category: "Kategorie",
             Item: "Artikel",
+            Tag: "Tag",
           }}
           typeNameUiSchemaOptionsMap={{
             Category: {
               dropdown: true,
             },
+            Tag: {
+              chips: true,
+            },
           }}
-  
+          uischemata={uischemata}
         >
           <App />
         </GraviolaProvider>
       </ThemeProvider>
     </QueryClientProvider>
   </StrictMode>,
-)
+);