+++ apps/testapp/CHANGELOG.md
@@ -1,5 +1,47 @@
 # testapp
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @graviola/edb-advanced-components@1.2.0
+  - @graviola/edb-core-types@1.1.3
+  - @graviola/edb-linked-data-renderer@1.1.6
+  - @graviola/semantic-json-form@1.1.3
+  - @graviola/edb-data-mapping@0.2.5
+  - @graviola/jsonld-utils@1.1.5
+  - @graviola/edb-state-hooks@1.2.5
+  - @graviola/entity-finder@1.1.5
+  - @graviola/rest-store-provider@1.1.7
+  - @graviola/semantic-jsonform-types@1.1.4
+  - @graviola/edb-basic-components@1.1.8
+
+## 0.0.4
+
+### Patch Changes
+
+- @graviola/rest-store-provider@1.1.6
+
+## 0.0.3
+
+### Patch Changes
+
+- make workspace depenedncies peer depenedncies
+- Updated dependencies
+  - @graviola/edb-linked-data-renderer@1.1.5
+  - @graviola/semantic-jsonform-types@1.1.4
+  - @graviola/edb-advanced-components@1.1.11
+  - @graviola/rest-store-provider@1.1.5
+  - @graviola/semantic-json-form@1.1.2
+  - @graviola/json-schema-utils@1.2.4
+  - @graviola/edb-basic-components@1.1.7
+  - @graviola/entity-finder@1.1.4
+  - @graviola/edb-data-mapping@0.2.4
+  - @graviola/jsonld-utils@1.1.4
+  - @graviola/edb-state-hooks@1.2.4
+  - @graviola/edb-core-types@1.1.2
+
 ## 0.0.2
 
 ### Patch Changes